+++ eRecruiter.Api/Responses/Policies/PolicyType.cs
@@ -1,0 +1,21 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace eRecruiter.Api.Responses.Policies
+{
+    public enum PolicyType
+    {
+        /// <summary>
+        /// Privacy policy
+        /// </summary>
+        PrivacyPolicy,
+
+        /// <summary>
+        /// Data Processing Agreement
+        /// </summary>
+        DataProcessingAgreement
+    }
+}